requirements.txt
Flask==2.3.3
line-bot-sdk==3.6.0
google-generativeai==0.3.0